added colours, need to make them work with user data now

+++ printf/src/conversion/num_conversions.c
@@ -6,13 +6,13 @@
 /*   By: dromansk <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/01/08 21:02:36 by dromansk          #+#    #+#             */
-/*   Updated: 2019/02/04 16:31:43 by dromansk         ###   ########.fr       */
+/*   Updated: 2019/02/05 14:09:28 by dromansk         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "printf.h"
 
-long	numlen(intmax_t n, int base)
+long	numlen(long long n, int base)
 {
 	long		i;
 
@@ -22,7 +22,7 @@ long	numlen(intmax_t n, int base)
 	return (i);
 }
 
-char	*ft_imaxtoa_base(intmax_t n, int base)
+char	*ft_imaxtoa_base(long long n, int base)
 {
 	char		*st;
 	intmax_t	v;
@@ -58,7 +58,7 @@ int		unumlen(unsigned long long value, unsigned int base)
 	return (i);
 }
 
-char	*ft_uimaxtoa_base(intmax_t n, int base)
+char	*ft_uimaxtoa_base(long long n, int base)
 {
 	char		*s;
 	uintmax_t	v;
